+++ src/openmp_benchmark/loopy-node-belief-propagation.c
@@ -7,7 +7,7 @@ int main(void) {
     fflush(out);
 
     struct joint_probability edge_joint_probability;
-    int dim_x, dim_y;
+    size_t dim_x, dim_y;
     set_joint_probability_yahoo_web(&edge_joint_probability, &dim_x, &dim_y);
 
     run_test_loopy_belief_propagation_mtx_files("/home/mjt5v/Desktop/belief-network-const-joint-probability/10_nodes_40_edges_2_beliefs.edges.mtx", "/home/mjt5v/Desktop/belief-network-const-joint-probability/10_nodes_40_edges_2_beliefs.nodes.mtx", &edge_joint_probability, dim_x, dim_y, out);
